refactor: split tunnel api clients
